Export libfuse private header directory
The Android FUSE daemon does not have permissions to mount().
This means we cannot call fuse_session_mount() because it calls
fuse_kern_mount() -> mount().

We can however add the opened /dev/fuse fd to the struct fuse_session.
Unfortunately, struct fuse_session is defined in fuse_i.h which is
in lib/ and not include/ so not exported.

This cl exports the headers in lib/

Change-Id: I0602fd8610f181caf7b5126a9bd2f98d2012dcf2