DO NOT MERGE FUSE passthrough: handle unstable kernel interface
The interface for FUSE passthrough is still under discussion upstream,
thus this change fixes the issue of being able to support multiple
interfaces for passthrough, depending on what the kernel provides.

This implementation tries to explore the different FUSE passthrough
interface implementation and updates an index to the one that has
been found working.

Bug: 174548608
Test: Manual read/write of passthrough files with extra printks
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I74e0e6c0691b37160a00af77fb18eda5342630f5