Extend initialization flags and FUSE_PASSTHROUGH to 64 bits
With the latest kernel version the FUSE_PASSTHROUGH flag now conflicts
with another flag.
As the flag bits in the original "flags" field are all covered, the
developer allocated another "flags2" field in the userspace <->
kernelspace communication.
While FUSE_PASSTHROUGH is kept as the last bit of either field, it's
userspace that needs to decide if it should be the last bit of "flags"
or "flags2" based on the FUSE version.

Test: launch_cvd (both android-mainline and android13-5.10) \
  `logcat FuseDaemon:V \*:S` shows no FUSE passthrough errors
Bug: 215310351
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Ib35e5e162b57a49fde1c36154adafb51172fc2bd