Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/433ebe82-6c55-4ff8-8670-4a0bf05e7e3c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If7893d14b1b51716c5ed514a52319d1e842ceeb6