Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/000576a9-7b05-4437-9142-f98fd46abe55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iddfb7018e23a7296a807b0bc5c4254ff60dcbeea