Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc2645ee-0fc9-4cc8-828c-f1507fe6059d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3d4adc808bb44e740330b6494850cb6093c025cb