Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/90009d39-c98a-4883-baa8-f2bed5aa8373). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8bcaf1e29478233cce8e68f7a1bb007693f4db0f