Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eded4e7c-e0b0-4dad-902c-4b65b49b4185). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6900a55c529436068abaab5a9a4df2a29a6ae90b