Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4bcea374-2dae-43e8-95a3-c182ab838070). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb720d2de847de4c8589f647194d294bd7c0db8d