Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5016f20e-051e-4674-83c1-b31804f628a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1cce4c0dc4c225f32959cb66ad860f81ac6e36c2