Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05b7ee69-9dae-4929-92de-06e09238233b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib9a2b2f5d04aae494fff343e73c520e516d0f20e