Revert "Workaround build breakage from full LTO optimization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48a0817c-4e6a-40fa-aa77-5cb8a3fe47ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c91fe9eba8c59d887f889b18b76a4dd3e44a728