Revert "Revert "Revert "Workaround build breakage from full LTO ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39d03588-4687-4e6d-aed6-32d4b46b8f0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib00cf8e32b8099db58b469107ac1b3c62f9794ba