Revert "Revert "Revert "Workaround build breakage from full LTO ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69b9aa60-7b29-43f6-93cc-d789661ba111). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26c3003e9ec89985acb40e86af8594fc69489364