Revert "Revert "Revert "Workaround build breakage from full LTO ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82e46832-28dd-49ee-af15-451053fa920e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf92552647b355399e1ea15ee3dcc7da72d01386